Rename raycast files.
