+++ Makefile
@@ -3,10 +3,10 @@
 #                                                         :::      ::::::::    #
 #    Makefile                                           :+:      :+:    :+:    #
 #                                                     +:+ +:+         +:+      #
-#    By: josantos <josantos@student.42.fr>          +#+  +:+       +#+         #
+#    By: msousa <mlrcbsousa@gmail.com>              +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2021/10/29 15:00:20 by msousa            #+#    #+#              #
-#    Updated: 2022/05/10 23:42:29 by josantos         ###   ########.fr        #
+#    Updated: 2022/05/11 00:18:52 by msousa           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -20,7 +20,7 @@ OBJ			= ${SRC:%.c=obj/%.o}
 SRC			= src/main.c src/draw.c src/hooks.c src/helpers.c src/test.c \
 					src/parse.c src/parse_1.c src/parse_2.c src/maplines.c src/map.c \
 					src/map_1.c src/settings.c src/elements.c src/elements_1.c \
-					src/maplines_1.c src/player.c src/raycast_1.c src/raycast_2.c \
+					src/maplines_1.c src/player.c src/raycast.c src/raycast_1.c \
 					src/move.c src/point.c src/validation.c src/vector.c
 
 NAME		= cub3d
